Bag System（TRUE）
When the player touching the items that can be picked up, it will be included in the backpack with special effects and sound effects (the sound effect has not been realized yet).
Note: before restarting the game, you need to reset the numerical parameters of Items( for 1) and Inventory( for 0).
The folder with same name  undering the home directory is uploaded incorrectly. Please delete it.